clase4/Practica/ml_services_comparison/client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["JARS29 <jandresrsalas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
grpcio = "^1.50.0"
grpcio-tools = "^1.50.0"
requests = "^2.30.0"
httpx = "^0.24.0" # Para el cliente GraphQL si se usa Async
numpy = "^1.23.0" 


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"